Add highlighting for parent <li>
Now, even if the actual event.target is a <cite> or <span> element inside of an <li> we always highlight the parent <li> element containing event.target. This will make it easier to highlight an <ol> or <ul> when I add that feature later on, since we'll know the direct parent of <li> will be one of these.