<a> uniqueness checks + append to localStorage
Wikiref now will append to existing data in  as opposed to overwriting it. Also, when checking the links associated with a particular <li>, a uniqueness check is done using the  property to ensure that there are no duplicate links for a given reference. Additionally, the logic that appends the reference selection options <div> to a given element has been generalized to account for the fact that  is not always equal to . Not adding this logic was rendering the <div> in weird places due to s that were placed after the last  node for a given <li> I was inspecting.