<a> uniqueness checks + append to localStorage
Wikiref now will append to existing data in localStorage as opposed to overwriting it. Also, when checking the links associated with a particular <li>, a uniqueness check is done using the href  property to ensure that there are no duplicate links for a given reference. Additionally, the logic that appends the reference selection options <div> to a given element has been generalized to account for the fact that element.lastElementChild  is not always equal to element.lastChild. Not adding this logic was rendering the <div> in weird places due to TextNodes that were placed after the last Element  node for a given <li> I was inspecting.