Add file download functionality
Now added functionality that enables downloading a JSON version of the references captured on a page. One current limitation of this is that it uses the current Wikipedia page name in order to name the downloaded JSON file; this is primarily because I kept running into errors when trying to enable the user the save the downloaded file with a given name. I may revisit this later if it becomes a big issue, but at the moment, I just wanted to implement the bare-minimum level of functionality.