Bugfix: Ref Selection Option Removal
This commit fixes a bug caused by the logic I'd originally implemented to remove the reference selection options <div> element. Essentially, the bug seems to have been that when a user would successfully extract one reference or group of references, re-enabling Selection Mode and then clicking on another reference(s) would cause the text of the previously extracted reference to be removed from the DOM. I believe this was caused by the fact that after selecting (but not yet confirming extraction of) a reference/group of references, there was logic added that would recursively travel up from the selected event.target to the closest parent <li>. The reason for this was to make expanding selection to the parent <ul> or <ol> of that <li> easier (if needed), as well as to make placement of the reference selection <div> consistent regardless of which part of a reference was clicked. After then confirming selection, the selection <div> that would've been attached to this <li> was removed. Whenever a new reference was selected, my original logic simply called element.removeChild(element.lastElementChild) to (erroneously) remove the reference selection <div> from the previously selected element. The unforeseen consequence of this, however, was that when a new reference/group of references was selected, the last child element of the previously selected <li>(or <ul> or <ol>) would no longer be the reference selection options <div> (because it was actually removed when selection was confirmed), but would instead likely be some part of the reference text itself.

To fix this problem, I've instead made it so that the document is searched for the id of the reference selection options <div>, and, if present, removed as a child of the previously selected element. Since we can only select one reference/group of references at a time, there should only ever be one of these elements in the DOM, which combined with conditional removal only if it is present means this should resolve the issue.