Implementation of "Delete References" popup
I used the storage.local API in order to store the status of a given page's reference status. The goal with this was to make it so that the extension popup on a given Wikipedia page will conditionally display a "Delete References" popup iff references have already been captured on that page.

This was done by storing a key-value pair for each visited Wikipedia page by using the page's URL as the key with the value being an object of the form { hidden: boolean }. This "hidden" field would indicate whether or not the "Delete References" <div> element should be hidden for the given page.

If you selected and successfully extracted some references, this would first trigger a saving of the references in the content script's localStorage, which would then update the "hidden" field to false for the object with key of location.href in storage.local. This would trigger a storage.onChanged event that would be handled by the background script, and the handler would update the popup by removing the "hidden" field from the classList of the "Delete References" <div>.

In the reverse direction, if a user clicked the "Delete References" <div> on the popup for a given page, the content script would first remove the references associated with location.href from localStorage, then would change the "hidden" field for the relevant object in storage.local to true. This update would propagate to the background script, which would update the popup by adding the "hidden" field to the classList of the "Delete References" <div>, thus hiding it.

Now, I'm not sure how, but it appears that I may have potentially introduced a bug while implementing this feature (I should probably make much smaller commits, but it was hard to make atomic changes in this case because I was so enthusiastic about seeing it fully work!). Specifically, now whenever a user selects and confirms extraction of a given reference/group of references, the cursor changes and the styling is removed, but it appears the modifySelectedRegionStyles event listener is still active on the document, as subsequent clicks will highlight references. I'll have to dig into why this is, but there's likely some small oversight I'm making.