Bugfix: Selection Mode Clean Exit
Fixed the issue where I was still in Selection Mode despite the fact that I had already confirmed selection of a reference. It turns out the solution was quite simple, like I'd originally thought: I neglected to ensure that the signature in removeListener matched with that of addListener whenever I removed the event listener attached to document.body in exitSelectionMode(). I literally just had to add the option true! :)