Part of editing feature: <p>'s replaced with <input>s
This is part of the implementation of the reference editing feature. Now, <p> elements in the <table> displaying ref data that are clicked are hidden and "replaced" by an <input> that contains the reference text that was in that <p> element.

Next step is to implement the actual editing flow, meaning that I might need to add:
- A way to specify that editing has been complete on the <input>.
- Finding the appropriate reference in localStorage and updating it after the 'change' event fires on the <input>, which I believe occurs after it is no longer focused.
- Removing the <input> from the DOM and redisplaying the <p>, this time with the text of the reference that was updated via the <input> element