Editing of references!
It appears that editing of references works in a fully functional way now, which is great!

The only problem I'm noticing now is that when I'm in editing mode, for some reason, trying to
download references no longer works. Here's what I think needs to happen next:
- When we enter "Edit" mode, highlight the pencil icon (invert the color palette). Reverse this
process if "Edit" icon is clicked again to get us back in the default mode.
- Add logic to disable edit mode if the "Download" button is complete. I may need to debug this a
bit, as I've noticed that currently, when I'm in "Edit" mode and click the "Download" button, I
receive an error about the content script being unable to connect to the background script.