Refactoring: background, popup, and content scripts
I realized I made a conceptual error when first creating this extension,
and what I'd labeled as the "background" script actually referred to the
popup script. This explains a lot of the weird errors I'd sometimes
encounter whenever I'd try to send a message to the "background" script
and received a message that the receiving end did not exist.

I must learn to strike a better balance between reading up on a
concept/domain beforehand and diving in and coding. Luckily, this was a
relatively easy fix in this case.

I divvied up the logic between the popup script and background script,
and renamed files appropriately so it's more clear what file belongs to
what part of extension.