Toggle Select Mode + Popup Updates Real-Time
This commmit contains logic that enables the user to remain in Select Mode after confirming selection of references.

Additionally, it contains updates to the UI that make it more responsive. It turns out the original logic that I'd implemented to do this for the "Delete References" object was actually incorrectly referencing the changes object that came from the storage.local.onChanged event. So, once I fixed that, the updates to the popup UI are essentially real-time! :)