Bugfix: Delete References now shows up only after ref extraction
In adding the logic for improving Select Mode, I noticed a bug! Essentially, when I'd first enter Select Mode, Delete References was popping up in the extension UI because I'd neglected to check the case that the "hidden" field in the storage.local object I was referring to and modifying throughout the codebase was undefined.